Refactor ContentView and DashboardView into smaller, manageable components
- Extract ContentView from 1,991 lines to 103 lines (95% reduction)
- Extract DashboardView from 872 lines to 224 lines
- Create organized file structure:
  * Core/Models: OnboardingEnums, OnboardingDataModel, MealModels
  * Core/Extensions: NotificationExtensions
  * Shared/Components: OnboardingSharedComponents
  * Features/Onboarding: OnboardingCoordinator, individual views
  * Features/Dashboard/Components: NutritionCard, GamificationSummaryCard
  * Features/Food: FoodScannerView
  * Features/Plan: PlanSummaryView
- Improve code maintainability and reusability
- Follow iOS architecture best practices with separation of concerns
