+++ FitApp/Core/Extensions/NotificationExtensions.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+// MARK: - Notification Names
+
+extension Notification.Name {
+    static let onboardingCompleted = Notification.Name("onboardingCompleted")
+}
