+++ FitApp/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,93 +1,12 @@
 {
   "images" : [
-    {
-      "idiom" : "iphone",
-      "scale" : "2x",
-      "size" : "20x20"
-    },
-    {
-      "idiom" : "iphone",
-      "scale" : "3x",
-      "size" : "20x20"
-    },
-    {
-      "idiom" : "iphone",
-      "scale" : "2x",
-      "size" : "29x29"
-    },
-    {
-      "idiom" : "iphone",
-      "scale" : "3x",
-      "size" : "29x29"
-    },
-    {
-      "idiom" : "iphone",
-      "scale" : "2x",
-      "size" : "40x40"
-    },
-    {
-      "idiom" : "iphone",
-      "scale" : "3x",
-      "size" : "40x40"
-    },
-    {
-      "idiom" : "iphone",
-      "scale" : "2x",
-      "size" : "60x60"
-    },
-    {
-      "idiom" : "iphone",
-      "scale" : "3x",
-      "size" : "60x60"
-    },
-    {
-      "idiom" : "ipad",
-      "scale" : "1x",
-      "size" : "20x20"
-    },
-    {
-      "idiom" : "ipad",
-      "scale" : "2x",
-      "size" : "20x20"
-    },
-    {
-      "idiom" : "ipad",
-      "scale" : "1x",
-      "size" : "29x29"
-    },
-    {
-      "idiom" : "ipad",
-      "scale" : "2x",
-      "size" : "29x29"
-    },
-    {
-      "idiom" : "ipad",
-      "scale" : "1x",
-      "size" : "40x40"
-    },
-    {
-      "idiom" : "ipad",
-      "scale" : "2x",
-      "size" : "40x40"
-    },
-    {
-      "idiom" : "ipad",
-      "scale" : "2x",
-      "size" : "76x76"
-    },
-    {
-      "idiom" : "ipad",
-      "scale" : "2x",
-      "size" : "83.5x83.5"
-    },
-    {
-      "idiom" : "ios-marketing",
-      "scale" : "1x",
-      "size" : "1024x1024"
-    }
+    { "size" : "60x60", "idiom" : "iphone", "filename" : "icon-60@2x-120.png", "scale" : "2x" },
+    { "size" : "60x60", "idiom" : "iphone", "filename" : "icon-60@3x-180.png", "scale" : "3x" },
+    { "size" : "40x40", "idiom" : "iphone", "filename" : "icon-40@2x-80.png",  "scale" : "2x" },
+    { "size" : "40x40", "idiom" : "iphone", "filename" : "icon-40@3x-120.png", "scale" : "3x" },
+    { "size" : "29x29", "idiom" : "iphone", "filename" : "icon-29@2x-58.png",  "scale" : "2x" },
+    { "size" : "29x29", "idiom" : "iphone", "filename" : "icon-29@3x-87.png",  "scale" : "3x" },
+    { "size" : "1024x1024", "idiom" : "ios-marketing", "filename" : "icon-ios-marketing-1024.png", "scale" : "1x" }
   ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
+  "info" : { "version" : 1, "author" : "xcode" }
 }
